update app to use bgg api dataset
