+++ src/app/app_wrangling.py
@@ -22,7 +22,7 @@ def call_boardgame_data():
     # files located in root
 
     boardgame_data = pd.read_csv(
-        "./data/app_data/new_tsne.csv",
+        "./data/app_data/bgg_data_tsne.csv",
         parse_dates=["year_published"],
         index_col=0,
     )
