+++ src/scripts/tsne_analysis.py
@@ -127,7 +127,7 @@ def tsne_analyse(onehot_df, user_df):
 
 if __name__ == "__main__":
     # load data and create subset for analysis
-    filename = "./data/app_data/new.csv"
+    filename = "./data/processed/bgg_wrangled.csv"
     # filename = "./data/app_data/board_game_date.csv"
     raw, mod = load_data(filename)
     print("Data loaded successfully")
@@ -142,5 +142,5 @@ if __name__ == "__main__":
 
     # save data
     combined_output = pd.concat([raw, result_cat, result_user], axis=1)
-    combined_output.to_csv("./data/app_data/new" + "_tsne.csv", index=False)
+    combined_output.to_csv("./data/app_data/bgg_data" + "_tsne.csv", index=False)
     print("Dataframe saved")
